# Kyber ASM ARMv7E-M/ARMv7-M: added assembly code