--- a/wolfcrypt/src/port/arm/thumb2-poly1305-asm_c.c
+++ b/wolfcrypt/src/port/arm/thumb2-poly1305-asm_c.c
@@ -93,17 +93,17 @@ void poly1305_blocks_thumb2_16(Poly1305* ctx, const byte* m, word32 len, int not
         "ADCS	r7, r7, r10\n\t"
         "ADD	%[m], %[m], #0x10\n\t"
         "ADC	r8, r8, r11\n\t"
-#ifdef WOLFSSL_SP_NO_UMAAL
+#ifdef WOLFSSL_ARM_ARCH_7M
         "STM	lr, {r4, r5, r6, r7, r8}\n\t"
 #else
         /* h[0]-h[2] in r4-r6 for multiplication. */
         "STR	r7, [lr, #12]\n\t"
         "STR	r8, [lr, #16]\n\t"
-#endif /* WOLFSSL_SP_NO_UMAAL */
+#endif /* WOLFSSL_ARM_ARCH_7M */
         "STR	%[m], [sp, #16]\n\t"
         "LDR	%[m], [sp, #12]\n\t"
         /* Multiply h by r */
-#ifdef WOLFSSL_SP_NO_UMAAL
+#ifdef WOLFSSL_ARM_ARCH_7M
         /* r0 = #0, r1 = r, lr = h, r2 = h[j], r3 = r[i] */
         "LDR	%[notLast], [%[m]]\n\t"
         "EOR	%[ctx], %[ctx], %[ctx]\n\t"
@@ -244,7 +244,7 @@ void poly1305_blocks_thumb2_16(Poly1305* ctx, const byte* m, word32 len, int not
         "UMAAL	r11, r12, %[notLast], r5\n\t"
         /* DONE */
         "LDM	sp, {r4, r5, r6}\n\t"
-#endif /* WOLFSSL_SP_NO_UMAAL */
+#endif /* WOLFSSL_ARM_ARCH_7M */
         /* r12 will be zero because r is masked. */
         /* Load length */
         "LDR	%[len], [sp, #20]\n\t"