--- a/wolfssl/wolfcrypt/wc_kyber.h
+++ b/wolfssl/wolfcrypt/wc_kyber.h
@@ -310,6 +310,22 @@ WOLFSSL_LOCAL int kyber_cmp_neon(const byte* a, const byte* b, int sz);
 WOLFSSL_LOCAL void kyber_csubq_neon(sword16* p);
 WOLFSSL_LOCAL void kyber_from_msg_neon(sword16* p, const byte* msg);
 WOLFSSL_LOCAL void kyber_to_msg_neon(byte* msg, sword16* p);
+#elif defined(__thumb__) && defined(WOLFSSL_ARMASM)
+#define kyber_ntt                   kyber_thumb2_ntt
+#define kyber_invntt                kyber_thumb2_invntt
+#define kyber_basemul_mont          kyber_thumb2_basemul_mont
+#define kyber_basemul_mont_add      kyber_thumb2_basemul_mont_add
+#define kyber_rej_uniform_c         kyber_thumb2_rej_uniform
+
+WOLFSSL_LOCAL void kyber_thumb2_ntt(sword16* r);
+WOLFSSL_LOCAL void kyber_thumb2_invntt(sword16* r);
+WOLFSSL_LOCAL void kyber_thumb2_basemul_mont(sword16* r, const sword16* a,
+     const sword16* b);
+WOLFSSL_LOCAL void kyber_thumb2_basemul_mont_add(sword16* r, const sword16* a,
+     const sword16* b);
+WOLFSSL_LOCAL void kyber_thumb2_csubq(sword16* p);
+WOLFSSL_LOCAL unsigned int kyber_thumb2_rej_uniform(sword16* p,
+    unsigned int len, const byte* r, unsigned int rLen);
 #endif
 
 #ifdef __cplusplus