… wolfcrypt/src/port/arm/thumb2-kyber-asm.S
… wolfcrypt/src/port/arm/thumb2-kyber-asm_c.c